Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,12 +1,12 @@
 #include <stdio.h>
 int main(){
-    int a;
+    float a;
     scanf("%d",&a);
 
-    int b;
+    float b;
     scanf("%d",&b);
 
-    int c;
+    float c;
     scanf("%d",&c);
 
     float average = (a+b+c)/3;
